+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version X.Y.Z (YYYYMMDDNN)
 
 - Display account creation date on dry-run page if a user never logged in (thanks to @35grain !)
+- Add support for AJAX-backed autocomplete elements in instance settings form
 
 
 ## Version 2.2.0 (2026051800)